requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.4
Django==3.2.9
django-ckeditor==6.4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.12.0
django-js-asset==2.0.0
django-jsonfield==1.4.1
django-storages==1.12.3
djangorestframework==3.13.1
drf-api-logger==1.1.10
idna==3.3
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
RandomWords==0.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.1.0
